Remove unused mocks and fixtures after refactoring
