--- __fixtures__/deletedFiles.js
@@ -1,55 +1,0 @@
-export default [
-  {
-    name: '/trash-bin/alice',
-    type: 'dir',
-    fileInfo: {
-      '{DAV:}resourcetype': ['{DAV:}collection']
-    },
-    tusSupport: null
-  },
-  {
-    name: '/trash-bin/alice/3875',
-    type: 'dir',
-    fileInfo: {
-      '{http://owncloud.org/ns}trashbin-original-filename': '.hidden-folder',
-      '{http://owncloud.org/ns}trashbin-original-location': '.hidden-folder',
-      '{http://owncloud.org/ns}trashbin-delete-datetime': 'Mon, 14 Jun 2021 12:15:33 GMT',
-      '{DAV:}resourcetype': ['{DAV:}collection']
-    },
-    tusSupport: null
-  },
-  {
-    name: '/trash-bin/alice/3871',
-    type: 'dir',
-    fileInfo: {
-      '{http://owncloud.org/ns}trashbin-original-filename': 'Documents',
-      '{http://owncloud.org/ns}trashbin-original-location': 'Documents',
-      '{http://owncloud.org/ns}trashbin-delete-datetime': 'Mon, 14 Jun 2021 12:15:33 GMT',
-      '{DAV:}resourcetype': ['{DAV:}collection']
-    },
-    tusSupport: null
-  },
-  {
-    name: '/trash-bin/alice/3873',
-    type: 'dir',
-    fileInfo: {
-      '{http://owncloud.org/ns}trashbin-original-filename': 'Photos',
-      '{http://owncloud.org/ns}trashbin-original-location': 'Photos',
-      '{http://owncloud.org/ns}trashbin-delete-datetime': 'Mon, 14 Jun 2021 12:15:33 GMT',
-      '{DAV:}resourcetype': ['{DAV:}collection']
-    },
-    tusSupport: null
-  },
-  {
-    name: '/trash-bin/alice/3852',
-    type: 'file',
-    fileInfo: {
-      '{http://owncloud.org/ns}trashbin-original-filename': 'ownCloud Manual.pdf',
-      '{http://owncloud.org/ns}trashbin-original-location': 'ownCloud Manual.pdf',
-      '{http://owncloud.org/ns}trashbin-delete-datetime': 'Wed, 09 Jun 2021 12:13:46 GMT',
-      '{DAV:}getcontentlength': '0',
-      '{DAV:}resourcetype': ''
-    },
-    tusSupport: null
-  }
-]
