--- __fixtures__/files.js
@@ -1,116 +1,0 @@
-export default {
-  '/': [
-    {
-      name: '',
-      type: 'dir',
-      fileInfo: {
-        '{http://owncloud.org/ns}permissions': 'RDNVCK',
-        '{http://owncloud.org/ns}favorite': '0',
-        '{http://owncloud.org/ns}fileid': '3861',
-        '{http://owncloud.org/ns}owner-id': 'alice',
-        '{http://owncloud.org/ns}owner-display-name': 'alice',
-        '{http://owncloud.org/ns}share-types': '',
-        '{http://owncloud.org/ns}privatelink': 'http://host.docker.internal:8080/f/3861',
-        '{http://owncloud.org/ns}size': '7546347',
-        '{DAV:}getlastmodified': 'Mon, 14 Jun 2021 09:41:16 GMT',
-        '{DAV:}getetag': '"60c7243c2e7f1"',
-        '{DAV:}resourcetype': ['{DAV:}collection']
-      },
-      tusSupport: null
-    },
-    {
-      name: '/.hidden-folder',
-      type: 'dir',
-      fileInfo: {
-        '{http://owncloud.org/ns}permissions': 'RDNVCK',
-        '{http://owncloud.org/ns}favorite': '0',
-        '{http://owncloud.org/ns}fileid': '14',
-        '{http://owncloud.org/ns}owner-id': 'admin',
-        '{http://owncloud.org/ns}owner-display-name': 'admin',
-        '{http://owncloud.org/ns}share-types': '',
-        '{http://owncloud.org/ns}privatelink': 'http://localhost:8080/f/14',
-        '{http://owncloud.org/ns}size': '0',
-        '{DAV:}getlastmodified': 'Thu, 24 Jun 2021 15:03:58 GMT',
-        '{DAV:}getetag': '"60d49edf01ab8"',
-        '{DAV:}resourcetype': ['{DAV:}collection']
-      },
-      tusSupport: null
-    },
-    {
-      name: '/Documents',
-      type: 'dir',
-      fileInfo: {
-        '{http://owncloud.org/ns}permissions': 'RDNVCK',
-        '{http://owncloud.org/ns}favorite': '0',
-        '{http://owncloud.org/ns}fileid': '3867',
-        '{http://owncloud.org/ns}owner-id': 'alice',
-        '{http://owncloud.org/ns}owner-display-name': 'alice',
-        '{http://owncloud.org/ns}share-types': '',
-        '{http://owncloud.org/ns}privatelink': 'http://host.docker.internal:8080/f/3867',
-        '{http://owncloud.org/ns}size': '36227',
-        '{DAV:}getlastmodified': 'Mon, 14 Jun 2021 09:41:16 GMT',
-        '{DAV:}getetag': '"60c7243c2e7f1"',
-        '{DAV:}resourcetype': ['{DAV:}collection']
-      },
-      tusSupport: null
-    },
-    {
-      name: '/Photos',
-      type: 'dir',
-      fileInfo: {
-        '{http://owncloud.org/ns}permissions': 'RDNVCK',
-        '{http://owncloud.org/ns}favorite': '0',
-        '{http://owncloud.org/ns}fileid': '3863',
-        '{http://owncloud.org/ns}owner-id': 'alice',
-        '{http://owncloud.org/ns}owner-display-name': 'alice',
-        '{http://owncloud.org/ns}share-types': '',
-        '{http://owncloud.org/ns}privatelink': 'http://host.docker.internal:8080/f/3863',
-        '{http://owncloud.org/ns}size': '1011464',
-        '{DAV:}getlastmodified': 'Mon, 14 Jun 2021 09:41:15 GMT',
-        '{DAV:}getetag': '"60c7243befd5d"',
-        '{DAV:}resourcetype': ['{DAV:}collection']
-      },
-      tusSupport: null
-    },
-    {
-      name: '/ownCloud Manual.pdf',
-      type: 'file',
-      fileInfo: {
-        '{http://owncloud.org/ns}permissions': 'RDNVW',
-        '{http://owncloud.org/ns}favorite': '0',
-        '{http://owncloud.org/ns}fileid': '3862',
-        '{http://owncloud.org/ns}owner-id': 'alice',
-        '{http://owncloud.org/ns}owner-display-name': 'alice',
-        '{http://owncloud.org/ns}share-types': '',
-        '{http://owncloud.org/ns}privatelink': 'http://host.docker.internal:8080/f/3862',
-        '{DAV:}getcontentlength': '6498656',
-        '{http://owncloud.org/ns}size': '6498656',
-        '{DAV:}getlastmodified': 'Mon, 14 Jun 2021 09:41:15 GMT',
-        '{DAV:}getetag': '"1f0ac81c04bc55c5d52a59f510c785fd"',
-        '{DAV:}resourcetype': '',
-        '{http://owncloud.org/ns}downloadURL': ''
-      },
-      tusSupport: null
-    }
-  ],
-  '/Documents': [
-    {
-      name: '/Documents',
-      type: 'dir',
-      fileInfo: {
-        '{http://owncloud.org/ns}permissions': 'RDNVCK',
-        '{http://owncloud.org/ns}favorite': '0',
-        '{http://owncloud.org/ns}fileid': '3867',
-        '{http://owncloud.org/ns}owner-id': 'alice',
-        '{http://owncloud.org/ns}owner-display-name': 'alice',
-        '{http://owncloud.org/ns}share-types': '',
-        '{http://owncloud.org/ns}privatelink': 'http://host.docker.internal:8080/f/3867',
-        '{http://owncloud.org/ns}size': '36227',
-        '{DAV:}getlastmodified': 'Mon, 14 Jun 2021 09:41:16 GMT',
-        '{DAV:}getetag': '"60c7243c2e7f1"',
-        '{DAV:}resourcetype': ['{DAV:}collection']
-      },
-      tusSupport: null
-    }
-  ]
-}
