--- __fixtures__/recipients.js
@@ -1,3 +1,0 @@
-export default {
-  users: [{ label: 'bob', value: { shareType: 0, shareWith: 'bob', userType: 0 } }]
-}
