--- __fixtures__/sharedWithMeFiles.js
@@ -1,134 +1,0 @@
-export default [
-  {
-    shareInfo: {
-      id: '22',
-      share_type: 0,
-      uid_owner: 'bob',
-      displayname_owner: 'bob',
-      permissions: 17,
-      stime: 1620810871,
-      parent: null,
-      expiration: null,
-      token: null,
-      uid_file_owner: 'bob',
-      displayname_file_owner: 'bob',
-      additional_info_owner: null,
-      additional_info_file_owner: null,
-      state: 2,
-      path: '/Documents',
-      mimetype: 'httpd/unix-directory',
-      storage_id: 'home::bob',
-      storage: 33,
-      item_type: 'folder',
-      item_source: 2243,
-      file_source: 2243,
-      file_parent: 1322,
-      file_target: '/Documents',
-      share_with: 'admin',
-      share_with_displayname: 'admin',
-      share_with_additional_info: null,
-      mail_send: 0,
-      attributes: null,
-      tags: []
-    }
-  },
-  {
-    shareInfo: {
-      id: '33',
-      share_type: 0,
-      uid_owner: 'bob',
-      displayname_owner: 'bob',
-      permissions: 17,
-      stime: 1620810871,
-      parent: null,
-      expiration: null,
-      token: null,
-      uid_file_owner: 'bob',
-      displayname_file_owner: 'bob',
-      additional_info_owner: null,
-      additional_info_file_owner: null,
-      state: 2,
-      path: '/.hidden-folder',
-      mimetype: 'httpd/unix-directory',
-      storage_id: 'home::bob',
-      storage: 33,
-      item_type: 'folder',
-      item_source: 2243,
-      file_source: 2243,
-      file_parent: 1322,
-      file_target: '/.hidden-folder',
-      share_with: 'admin',
-      share_with_displayname: 'admin',
-      share_with_additional_info: null,
-      mail_send: 0,
-      attributes: null,
-      tags: []
-    }
-  },
-  {
-    shareInfo: {
-      id: '23',
-      share_type: 0,
-      uid_owner: 'bob',
-      displayname_owner: 'bob',
-      permissions: 17,
-      stime: 1620810871,
-      parent: null,
-      expiration: null,
-      token: null,
-      uid_file_owner: 'bob',
-      displayname_file_owner: 'bob',
-      additional_info_owner: null,
-      additional_info_file_owner: null,
-      state: 2,
-      path: '/Photos',
-      mimetype: 'httpd/unix-directory',
-      storage_id: 'home::bob',
-      storage: 33,
-      item_type: 'folder',
-      item_source: 2244,
-      file_source: 2244,
-      file_parent: 1322,
-      file_target: '/Photos',
-      share_with: 'admin',
-      share_with_displayname: 'admin',
-      share_with_additional_info: null,
-      mail_send: 0,
-      attributes: null,
-      tags: []
-    }
-  },
-  {
-    shareInfo: {
-      id: '24',
-      share_type: 0,
-      uid_owner: 'bob',
-      displayname_owner: 'bob',
-      permissions: 19,
-      stime: 1623672189,
-      parent: null,
-      expiration: null,
-      token: null,
-      uid_file_owner: 'bob',
-      displayname_file_owner: 'bob',
-      additional_info_owner: null,
-      additional_info_file_owner: null,
-      state: 1,
-      path: '/ownCloud Manual.pdf',
-      mimetype: 'text/plain',
-      storage_id: 'home::bob',
-      storage: 33,
-      item_type: 'file',
-      item_source: 2248,
-      file_source: 2248,
-      file_parent: 1322,
-      file_target: '/ownCloud Manual.pdf',
-      share_with: 'admin',
-      share_with_displayname: 'admin',
-      share_with_additional_info: null,
-      mail_send: 0,
-      attributes: null,
-      tags: []
-    }
-  }
-]
