--- __fixtures__/users.js
@@ -1,36 +1,0 @@
-export default {
-  alice: {
-    quota: {
-      free: 37883293696,
-      used: 7546347,
-      total: 37890840043,
-      relative: 0.02,
-      definition: 'default'
-    },
-    email: {},
-    id: 'alice',
-    name: 'alice',
-    displayname: 'Alice Hansen',
-    additionalInfo: 'alice@example.org',
-    home: '/mnt/data/files/alice',
-    two_factor_auth_enabled: 'false',
-    uuid: 1
-  },
-  admin: {
-    quota: {
-      free: 37883293696,
-      used: 7546347,
-      total: 37890840043,
-      relative: 0.02,
-      definition: 'default'
-    },
-    email: {},
-    id: 'admin',
-    name: 'admin',
-    displayname: 'admin',
-    additionalInfo: 'admin@example.org',
-    home: '/mnt/data/files/admin',
-    two_factor_auth_enabled: 'false',
-    uuid: 2
-  }
-}
